refactor: move shared objects into building blocks
