+++ Backend/src/BuildingBlocks/BuildingBlocks.Domain/Abstractions/Entity.cs
@@ -1,12 +1,8 @@
-﻿namespace BuildingBlocks.Domain
+﻿namespace BuildingBlocks.Domain.Abstractions
 {
-    public interface IEntity<out T> : IEntity
+    public abstract class Entity<T> : IEntity<T>
     {
-        T Id { get; }
-    }
-
-    public interface IEntity
-    {
+        public required T Id { get; init; }
         public DateTime? CreatedAt { get; set; }
         public string? CreatedBy { get; set; }
         public DateTime? LastModifiedAt { get; set; }
