+++ Backend/src/BuildingBlocks/BuildingBlocks.Domain/GlobalUsing.cs
@@ -1,0 +1,1 @@
+global using MediatR;
